Double usage for more accuracy

+++ Calculator/INode.cs
@@ -8,6 +8,6 @@ namespace Calculator
 {
     public interface INode
     {
-        float GetResult();
+        double GetResult();
     }
 }
